creates billing router for end points, revises balance display information, balance updates on successful payment for 1 or 12 jobs
